feat(rust): 生产对齐 webadmin 全页面 — me/users/gateway/providers/models/auth/brand/portal/channels/connectors/audit-settings/usage(provider归并)/requests/report-subscriptions/server-info/balance/skills/agents/capabilities-approvals(org过滤) + SPA fallback修复 + 时区/类型修复

+++ rust/crates/dsh-store/src/groups.rs
@@ -110,9 +110,9 @@ pub async fn user_groups_batch(
     }
     let rows = q.fetch_all(pool).await.map_err(map_db_error)?;
     for r in rows {
-        let uid: i64 = r.get("user_id");
+        let uid: i32 = r.get("user_id");
         let name: String = r.get("name");
-        out.entry(uid).or_default().push(name);
+        out.entry(uid as i64).or_default().push(name);
     }
     Ok(out)
 }
